+++ java/alix/lucene/search/ScorerBM25.java
@@ -56,7 +56,7 @@ public class ScorerBM25 extends Scorer
   @Override
   public void weight(final long occsMatch, final int docsMatch)
   {
-    this.idf = (double) Math.log(1.0 + (docsAll - docsMatch + 0.5D) / (docsMatch + 0.5D));
+    this.idf = Math.log(1.0 + (docsAll - docsMatch + 0.5D) / (docsMatch + 0.5D));
   }
 
   @Override
